% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way.R
\name{aggregate_one_way}
\alias{aggregate_one_way}
\title{Aggregate data for a one-way plot}
\usage{
aggregate_one_way(
  dt,
  var,
  obs,
  expo_col,
  split,
  bins,
  type_agg,
  time_unit = NA_character_
)
}
\value{
data.table with columns: x, split, \if{html}{\out{<obs columns>}}, exposure
}
\description{
Aggregate data for a one-way plot
}
\keyword{internal}
