% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp.R
\name{aggregate_pdp_oneway}
\alias{aggregate_pdp_oneway}
\title{Aggregate observed and in-sample predicted values per bin}
\usage{
aggregate_pdp_oneway(dt, obs, expo_col)
}
\value{
data.table with columns: .bin, obs_mean, pred_mean, exposure
}
\description{
Operates on \code{dt} which must already have columns \code{.bin}, \code{.pred},
the obs column, and the exposure column.
}
\keyword{internal}
