% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{assert_col_exists}
\alias{assert_col_exists}
\title{Abort if any of \code{cols} are missing from \code{data}}
\usage{
assert_col_exists(data, cols, arg_name)
}
\description{
Abort if any of \code{cols} are missing from \code{data}
}
\keyword{internal}
