% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_equal_exposure}
\alias{bin_equal_exposure}
\title{Bin a sorted numeric vector into n equal-exposure groups}
\usage{
bin_equal_exposure(x_sorted, bins)
}
\description{
Bin a sorted numeric vector into n equal-exposure groups
}
\keyword{internal}
