% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_equal_range}
\alias{bin_equal_range}
\title{Bin a numeric vector into n equal-range groups}
\usage{
bin_equal_range(x, bins)
}
\description{
Bin a numeric vector into n equal-range groups
}
\keyword{internal}
