% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp.R
\name{coerce_to_original}
\alias{coerce_to_original}
\title{Coerce a bin label back to the original column type}
\usage{
coerce_to_original(original_vec, label)
}
\description{
Coerce a bin label back to the original column type
}
\keyword{internal}
