% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{distribution.modelblueprint}
\alias{distribution.modelblueprint}
\title{Target distribution for a modelblueprint}
\usage{
\method{distribution}{modelblueprint}(
  data,
  set = c("train", "test", "holdout"),
  split = NA_character_,
  bins = 35L,
  ...
)
}
\arguments{
\item{data}{A \code{modelblueprint}.}

\item{set}{\verb{[character(1)]} Which dataset to use: \code{"train"}, \code{"test"},
or \code{"holdout"}. Default \code{"train"}.}

\item{split}{\verb{[character(1) | NA]} Optional split variable.}

\item{bins}{\verb{[integer(1)]} Number of bins. Default \code{35L}.}

\item{...}{Further arguments passed to \code{\link[=distribution]{distribution()}}.}
}
\value{
A plotly object or data.table depending on \code{ret}.
}
\description{
Calls \code{\link[=distribution]{distribution()}} using the modelblueprint's target and exposure slots,
showing the exposure-weighted distribution of the target variable.
}
