% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_metadata}
\alias{extract_metadata}
\alias{extract_target}
\alias{extract_yhat_name}
\alias{extract_exposure_name}
\alias{extract_exposure_value}
\alias{extract_exposure_zero_rep}
\alias{extract_offset_name}
\alias{extract_offset_value}
\alias{extract_display_name}
\alias{extract_deploy_notes}
\title{Extract scalar metadata from a modelblueprint}
\usage{
extract_target(x, ...)

extract_yhat_name(x, ...)

extract_exposure_name(x, ...)

extract_exposure_value(x, ...)

extract_exposure_zero_rep(x, ...)

extract_offset_name(x, ...)

extract_offset_value(x, ...)

extract_display_name(x, ...)

extract_deploy_notes(x, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{...}{Unused.}
}
\value{
The slot value (\code{NA} when not set).
}
\description{
Extract scalar metadata from a modelblueprint
}
