% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_pipeline}
\alias{extract_pipeline}
\alias{extract_pre_process_fun}
\alias{extract_feat_eng_fun}
\alias{extract_post_process_fun}
\title{Extract pipeline functions from a modelblueprint}
\usage{
extract_pre_process_fun(x, ...)

extract_feat_eng_fun(x, ...)

extract_post_process_fun(x, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{...}{Unused.}
}
\value{
The pipeline function.
}
\description{
Extract pipeline functions from a modelblueprint
}
\seealso{
\code{\link[=set_pre_process_fun]{set_pre_process_fun()}}, \code{\link[=set_feat_eng_fun]{set_feat_eng_fun()}}, \code{\link[=set_post_process_fun]{set_post_process_fun()}}
}
