% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_splits}
\alias{extract_splits}
\alias{extract_train}
\alias{extract_test}
\alias{extract_holdout}
\title{Extract a data split from a modelblueprint}
\usage{
extract_train(x, ...)

extract_test(x, ...)

extract_holdout(x, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{...}{Unused.}
}
\value{
A \code{data.frame} / \code{data.table}, or \code{NULL} if the split was not set.
}
\description{
Extract a data split from a modelblueprint
}
\seealso{
\code{\link[=set_train]{set_train()}}, \code{\link[=set_test]{set_test()}}, \code{\link[=set_holdout]{set_holdout()}}
}
