% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{filter.modelblueprint}
\alias{filter.modelblueprint}
\title{Filter rows in a modelblueprint's datasets}
\usage{
\method{filter}{modelblueprint}(.data, ..., sets = c("train", "test", "holdout"))
}
\arguments{
\item{.data}{A \code{modelblueprint}.}

\item{...}{Filter expressions passed to \code{dplyr::filter()}.}

\item{sets}{Which datasets to filter. Default: all non-NULL.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Filter rows in a modelblueprint's datasets
}
