% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{loadMB}
\alias{loadMB}
\title{Load a modelblueprint from disk}
\usage{
loadMB(path)
}
\arguments{
\item{path}{Path to the \code{.tar.gz} archive created by \code{\link[=saveMB]{saveMB()}}.}
}
\value{
A fully reconstructed \code{modelblueprint} object.
}
\description{
Reconstructs a \code{modelblueprint} from a \code{.tar.gz} archive created by \code{\link[=saveMB]{saveMB()}}.
}
\seealso{
\code{\link[=saveMB]{saveMB()}}
}
