% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way.R
\name{make_palette}
\alias{make_palette}
\title{Make a colour palette of length n, handling n < 3 gracefully}
\usage{
make_palette(n, palette = "Paired")
}
\description{
Make a colour palette of length n, handling n < 3 gracefully
}
\keyword{internal}
