% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{mutate.modelblueprint}
\alias{mutate.modelblueprint}
\title{Mutate columns in a modelblueprint's datasets}
\usage{
\method{mutate}{modelblueprint}(.data, ..., sets = c("train", "test", "holdout"))
}
\arguments{
\item{.data}{A \code{modelblueprint}.}

\item{...}{Expressions passed to \code{dplyr::mutate()}.}

\item{sets}{Which datasets to mutate. Default: all non-NULL.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Mutate columns in a modelblueprint's datasets
}
