% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{predict.modelblueprint}
\alias{predict.modelblueprint}
\title{Generate predictions from a modelblueprint}
\usage{
\method{predict}{modelblueprint}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{modelblueprint}.}

\item{newdata}{A \code{data.frame} or \code{data.table}.}

\item{...}{Unused.}
}
\value{
A numeric vector of predictions.
}
\description{
Applies \code{pre_process_fun} -> \code{feat_eng_fun} -> model prediction ->
\code{post_process_fun} to \code{newdata}.
}
