% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{set_features}
\alias{set_features}
\alias{set_original_inputs}
\alias{set_feature_names}
\title{Replace feature name vectors in a modelblueprint}
\usage{
set_original_inputs(x, value, ...)

set_feature_names(x, value, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{value}{A character vector of column names.}

\item{...}{Unused.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Replace feature name vectors in a modelblueprint
}
\seealso{
\code{\link[=extract_original_inputs]{extract_original_inputs()}}, \code{\link[=extract_feature_names]{extract_feature_names()}}
}
