% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{set_metadata}
\alias{set_metadata}
\alias{set_target}
\alias{set_yhat_name}
\alias{set_exposure_name}
\alias{set_exposure_value}
\alias{set_exposure_zero_rep}
\alias{set_offset_name}
\alias{set_offset_value}
\alias{set_display_name}
\alias{set_deploy_notes}
\title{Set scalar metadata on a modelblueprint}
\usage{
set_target(x, value, ...)

set_yhat_name(x, value, ...)

set_exposure_name(x, value, ...)

set_exposure_value(x, value, ...)

set_exposure_zero_rep(x, value, ...)

set_offset_name(x, value, ...)

set_offset_value(x, value, ...)

set_display_name(x, value, ...)

set_deploy_notes(x, value, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{value}{A single string (or numeric for \code{set_exposure_value} /
\code{set_exposure_zero_rep} / \code{set_offset_value}).}

\item{...}{Unused.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Set scalar metadata on a modelblueprint
}
