% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{set_pipeline}
\alias{set_pipeline}
\alias{set_pre_process_fun}
\alias{set_feat_eng_fun}
\alias{set_post_process_fun}
\title{Replace pipeline functions in a modelblueprint}
\usage{
set_pre_process_fun(x, value, ...)

set_feat_eng_fun(x, value, ...)

set_post_process_fun(x, value, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{value}{A function with the expected signature for that pipeline stage.}

\item{...}{Unused.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Replace pipeline functions in a modelblueprint
}
\seealso{
\code{\link[=extract_pre_process_fun]{extract_pre_process_fun()}}, \code{\link[=extract_feat_eng_fun]{extract_feat_eng_fun()}},
\code{\link[=extract_post_process_fun]{extract_post_process_fun()}}
}
