% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{set_splits}
\alias{set_splits}
\alias{set_train}
\alias{set_test}
\alias{set_holdout}
\title{Replace a data split in a modelblueprint}
\usage{
set_train(x, value, ...)

set_test(x, value, ...)

set_holdout(x, value, ...)
}
\arguments{
\item{x}{A \code{modelblueprint}.}

\item{value}{A \code{data.frame} or \code{data.table}.}

\item{...}{Unused.}
}
\value{
A new \code{modelblueprint}.
}
\description{
Replace a data split in a modelblueprint
}
\seealso{
\code{\link[=extract_train]{extract_train()}}, \code{\link[=extract_test]{extract_test()}}, \code{\link[=extract_holdout]{extract_holdout()}}
}
