% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{filter.ModelBlueprint}
\alias{filter.ModelBlueprint}
\title{Filter rows in a ModelBlueprint's datasets}
\usage{
\method{filter}{ModelBlueprint}(.data, ..., sets = c("train", "test", "holdout"))
}
\arguments{
\item{.data}{A \code{ModelBlueprint}.}

\item{...}{Filter expressions passed to \code{dplyr::filter()}.}

\item{sets}{Which datasets to filter. Default: all non-NULL.}
}
\value{
A new \code{ModelBlueprint}.
}
\description{
Filter rows in a ModelBlueprint's datasets
}
