% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{left_join.ModelBlueprint}
\alias{left_join.ModelBlueprint}
\title{Left-join into a ModelBlueprint's datasets}
\usage{
\method{left_join}{ModelBlueprint}(x, y, by = NULL, ..., sets = c("train", "test", "holdout"))
}
\arguments{
\item{x}{A \code{ModelBlueprint}.}

\item{y}{A \code{data.frame} to join.}

\item{by}{Join keys.}

\item{...}{Passed to \code{dplyr::left_join()}.}

\item{sets}{Which datasets to join. Default: all non-NULL.}
}
\value{
A new \code{ModelBlueprint}.
}
\description{
Left-join into a ModelBlueprint's datasets
}
