% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{loadmb}
\alias{loadmb}
\alias{loadMB}
\title{Load a modelblueprint from disk}
\usage{
loadmb(path)

loadMB(path)
}
\arguments{
\item{path}{Path to the \code{.tar.gz} archive created by \code{\link[=savemb]{savemb()}}.}
}
\value{
A fully reconstructed \code{modelblueprint} object.
}
\description{
Reconstructs a \code{modelblueprint} from a \code{.tar.gz} archive created by \code{\link[=savemb]{savemb()}}.

\code{loadMB()} is deprecated; use \code{\link[=loadmb]{loadmb()}} instead.
}
\seealso{
\code{\link[=savemb]{savemb()}}
}
