% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBlueprint-package.R
\docType{package}
\name{modelblueprint-package}
\alias{modelblueprint-package}
\title{modelblueprint: One-Way and Partial Dependence Plots for Models}
\description{
Provides fast, production-ready one-way analysis plots and partial dependence plots (PDPs) designed for insurance and credit pricing workflows. One-way plots show exposure-weighted means of one or more observed variables across bins of a feature, with optional split grouping. PDP plots reveal the marginal effect of a feature on model output, controlling for all other features, and work with any model that implements a predict() method including H2O models. All plots use a consistent dual-axis Plotly style: yellow exposure bars on the left axis and weighted means on the right. Aggregation is handled by data.table and scales comfortably to datasets of 1-2 million rows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matt-or.github.io/modelblueprint/}
}

}
\author{
\strong{Maintainer}: Matt \email{fermoymatt@gmail.com}

Authors:
\itemize{
  \item Matt \email{fermoymatt@gmail.com}
}

}
\keyword{internal}
