% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelblueprint.R
\name{mutate.ModelBlueprint}
\alias{mutate.ModelBlueprint}
\title{Mutate columns in a ModelBlueprint's datasets}
\usage{
\method{mutate}{ModelBlueprint}(.data, ..., sets = c("train", "test", "holdout"))
}
\arguments{
\item{.data}{A \code{ModelBlueprint}.}

\item{...}{Expressions passed to \code{dplyr::mutate()}.}

\item{sets}{Which datasets to mutate. Default: all non-NULL.}
}
\value{
A new \code{ModelBlueprint}.
}
\description{
Mutate columns in a ModelBlueprint's datasets
}
