% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way.R
\name{smart_level_order}
\alias{smart_level_order}
\title{Sort x-axis labels: numerics / intervals first, then categoricals, NA last}
\usage{
smart_level_order(x)
}
\description{
Sort x-axis labels: numerics / intervals first, then categoricals, NA last
}
\keyword{internal}
